Fix Optional::iter_range
In the None case we only need to produce `range` number of values, not a
value for every item in the column.

In practice doesn't cause a serious issue, as these iterators are
usually zipped with another shorter iterator.
Still queries with only Options will see a slight speedup in parallel
iteration on archetypes without any queried component.